manejo de errores y mensajes al front

+++ controllers/productController.js
@@ -4,16 +4,17 @@ import fs from "fs-extra"
 import path from "path"
 
 const addProduct = async (req, res) => {
-  const { name, products, subTotal, profit, total } = JSON.parse(req.body.jsonData);
+  const { name, products, subTotal, percentage, profit, total } = JSON.parse(req.body.jsonData);
 
   if(name.trim() === "" || !Array.isArray(products) || products.length === 0){
     const error = new Error("Existen campos vacios o Se requiere al menos 1 producto");
     return res.status(400).json({ msg: error.message });
   }
- 
-  if(req.files && !req.files.image.mimetype.startsWith('image/')){
-    const error = new Error("El archivo subido no es una imagen válida");
-    return res.status(400).json({ msg: error.message });
+
+  const existingProduct = await Product.findOne({ name });
+  if (existingProduct) {
+     res.status(400).json({ msg: "La canasta ya existe", name });
+     return
   }
 
   const productData = {
@@ -21,13 +22,13 @@ const addProduct = async (req, res) => {
     seller: req.seller._id,
     products,
     subTotal,
+    percentage,
     profit,
     total
   };
   
   if (req.files?.image && req.files.image.mimetype.startsWith('image/')) {
-    const allowedExtensions = ['.jpg', '.jpeg', '.png', '.gif','.tiff', '.webp', '.svg',
-    '.ico','.avif']; 
+    const allowedExtensions = ['.jpg', '.jpeg', '.png', '.gif','.tiff', '.webp', '.svg', '.ico','.avif']; 
     const fileExtension = path.extname(req.files.image.name).toLowerCase();
     if (allowedExtensions.includes(fileExtension)) {
       const result = await uploadImage(req.files.image.tempFilePath);
@@ -36,7 +37,8 @@ const addProduct = async (req, res) => {
         secure_url: result.secure_url
       };
     } else {
-      res.status(400).json({ msg: "Archivo no válido. Solo se permiten imágenes jpge, jpg, png, gif, tiff, web, svg, ico, avif" });
+      res.status(400).json({ msg: "Imagen no válido" });
+      return 
     }
   }
 
@@ -44,41 +46,56 @@ const addProduct = async (req, res) => {
   try {
     const savedProduct = await product.save();
     res.json(savedProduct);
-    await fs.unlink(req.files.image.tempFilePath)
+    if(req.files?.image){
+      await fs.unlink(req.files.image.tempFilePath)
+    }
   } catch (error) {
-    console.log(error)
-    res.status(500).json({ msg: "Esta Canasta ya Existe"});
+    console.log(error, "error")
+    res.status(500).json({ msg: "Error al crear canasta"});
+    if(req.files?.image){
+      await deleteImage(product.image.public_id)
+      await fs.unlink(req.files.image.tempFilePath)
+    }
   }
 }; 
 
 const getProducts = async (req, res) => {
-  const product = await Product.find().where("seller").equals(req.seller);
-  res.json(product);
+  try {
+    const product = await Product.find().where("seller").equals(req.seller);
+    res.json(product);
+  } catch (error) {
+    console.log(error)
+    return res.status(500).json({ msg: "Error interno del servidor" });
+  }
 };
 
 const getProduct = async (req, res) => {
-  const { id } = req.params;
-  if (id.length !== 24) {
-    return res.status(403).json({ msg: "No Encontrado" });
-  }
-  const product = await Product.findById(id);
+  try {
+    const { id } = req.params;
+    if (id.length !== 24) {
+      return res.status(403).json({ msg: "No Encontrado" });
+    }
+    const product = await Product.findById(id);
 
-  if (!product) {
-    return res.status(404).json({ msg: "No Encontrado" });
-  }
+    if (!product) {
+      return res.status(404).json({ msg: "No Encontrado" });
+    }
 
-  if (product.seller._id.toString() !== req.seller._id.toString()){
-    return res.json({ msg: "Accion no valida" });
-  }
+    if (product.seller._id.toString() !== req.seller._id.toString()){
+      return res.json({ msg: "Accion no valida" });
+    }
 
-  res.json(product);
+    res.json(product);
+  } catch (error) {
+    console.error("Error en getProduct:", error);
+    return res.status(500).json({ msg: "Error interno del servidor" });
+  }
 };
 
 const updateProduct = async (req, res) => {
   const { id } = req.params;
-  const { name, subtotal, profit, total, percentage, products} = JSON.parse(req.body.jsonData);
-
-
+  const { name, subtotal, profit, total, percentage, products } = JSON.parse(req.body.jsonData);
+ 
   if(name.trim() === "" || !Array.isArray(products) || products.length === 0){
     const error = new Error("Existen campos vacios o Se requiere al menos 1 producto");
     return res.status(400).json({ msg: error.message });
@@ -98,23 +115,27 @@ const updateProduct = async (req, res) => {
     return res.json({ msg: "Accion no valida" });
   }
 
-  product.name = name || product.name;
-  
+  const existingProduct = await Product.findOne({ name });
+  if (existingProduct && existingProduct._id.toString() !== id) {
+    return res.status(400).json({ msg: "La canasta ya existe", name });
+  }
+
+  product.name = name || product.name
+  //image
   if (req.files?.image && req.files.image.mimetype.startsWith('image/')) {
     const fileExtension = path.extname(req.files.image.name).toLowerCase();
     const allowedExtensions = ['.jpg', '.jpeg', '.png', '.gif', '.tiff', '.webp', '.svg', '.ico', '.avif'];
-  
+
     if (!allowedExtensions.includes(fileExtension)) {
       return res.status(400).json({ msg: "Extensión de archivo no permitida" });
     }
   
-    try {
-      const result = await uploadImage(req.files.image.tempFilePath);
-  
+    try {  
       if (product.image?.public_id) {
         await deleteImage(product.image.public_id);
       }
-      
+
+      const result = await uploadImage(req.files.image.tempFilePath);
       product.image = {
         public_id: result.public_id,
         secure_url: result.secure_url,
@@ -124,11 +145,14 @@ const updateProduct = async (req, res) => {
     } catch (error) {
       return res.status(500).json({ msg: "Error al actualizar la imagen" });
     }
-  } else if(product.image.public_id){
+  } 
+
+  if (!req.files && !req.body.image.startsWith("https") && product.image?.public_id) {
     await deleteImage(product.image.public_id);
-    product.image = null
+    product.image = null;
   }
 
+  //products
   if (products && Array.isArray(products)) {
     product.products = products.map(updatedProduct => {
       const existingProduct = product.products.find(p => p._id.equals(updatedProduct._id));
@@ -140,23 +164,23 @@ const updateProduct = async (req, res) => {
       }
       return updatedProduct;
     });
-  }else{
+  } else {
     res.status(500).json({ msg: "Error en products"});
   }
-
-  product.subtotal = subtotal || product.subtotal;
-  product.percentage = percentage || product.percentage
-  product.profit = profit || product.profit;
-  product.total = total || product.total;
+ 
+  product.subtotal = (percentage !== undefined) ? subtotal : product.subtotal;
+  product.percentage = (percentage !== undefined) ? percentage : product.percentage;
+  product.profit = (profit !== undefined) ? profit : product.profit;
+  product.total = (total !== undefined) ? total : product.total;
 
   try {
     const updatedProduct = await product.save();
     res.json(updatedProduct);
   } catch (error) {
     console.log(error)
-    res.status(500).json({ msg: "Esta Canasta ya Existe"});
+    res.status(500).json({ msg: "Error al guardar la canasta"});
   }
-}; 
+};
 
 const deleteProduct = async (req, res) => {
   const { id } = req.params;
@@ -187,9 +211,3 @@ const deleteProduct = async (req, res) => {
 };
 
 export { addProduct, getProducts, getProduct, updateProduct, deleteProduct };
-
-/* if(!req.files?.image){
-    await deleteImage(product.image.public_id);
-    product.image = undefined
-  }
-*/
